adds string interpretation to 3d
all the stuff is copied over from calculator2d, which is now commented out.
