+++ Calculator3D.cs
@@ -3,9 +3,9 @@ using System;
 
 public class Calculator3D : Spatial
 {
-    public void sig_GetCameraPosition()
-    {
-        GetNode<Gridlines>("Gridlines").GlobalCameraOrigin = 
-                GetNode<Camera>("CameraPivot/Camera").GlobalTransform.origin;
-    }
+	public void sig_GetCameraPosition()
+	{
+		GetNode<Gridlines>("Gridlines").GlobalCameraOrigin = 
+				GetNode<Camera>("CameraPivot/Camera").GlobalTransform.origin;
+	}
 }
